Change u64 to usize in vmm/devices/virtio/queue
We can use usize in place of u64 here to reduce type conversion
warnings. On 64-bit systems the types are equivalent, but Clippy allows
automatic conversions in the one direction but not in the other.

+++ src/vmm/src/devices/virtio/queue.rs
@@ -238,7 +238,7 @@ impl Queue {
 
     /// Validates the queue's in-memory layout is correct.
     pub fn is_layout_valid(&self, mem: &GuestMemoryMmap) -> bool {
-        let queue_size = u64::from(self.actual_size());
+        let queue_size = usize::from(self.actual_size());
         let desc_table = self.desc_table;
         let desc_table_size = 16 * queue_size;
         let avail_ring = self.avail_ring;
@@ -263,21 +263,21 @@ impl Queue {
             error!("virtio queue used ring breaks alignment constraints");
             false
         // range check entire descriptor table to be assigned valid guest physical addresses
-        } else if mem.get_slice(desc_table, desc_table_size as usize).is_err() {
+        } else if mem.get_slice(desc_table, desc_table_size).is_err() {
             error!(
                 "virtio queue descriptor table goes out of bounds: start:0x{:08x} size:0x{:08x}",
                 desc_table.raw_value(),
                 desc_table_size
             );
             false
-        } else if mem.get_slice(avail_ring, avail_ring_size as usize).is_err() {
+        } else if mem.get_slice(avail_ring, avail_ring_size).is_err() {
             error!(
                 "virtio queue available ring goes out of bounds: start:0x{:08x} size:0x{:08x}",
                 avail_ring.raw_value(),
                 avail_ring_size
             );
             false
-        } else if mem.get_slice(used_ring, used_ring_size as usize).is_err() {
+        } else if mem.get_slice(used_ring, used_ring_size).is_err() {
             error!(
                 "virtio queue used ring goes out of bounds: start:0x{:08x} size:0x{:08x}",
                 used_ring.raw_value(),
